+++ src/compiler/codegen/src/lib.rs
@@ -1,3 +1,5 @@
+//! Transpiles the [CloesceIdl] into backend and frontend APIs.
+
 pub mod backend;
 pub mod client;
 pub mod wrangler;
